requirements.txt
Django==5.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
django-filter==23.5
scikit-learn==1.4.0
numpy==1.26.3
markdown==3.5.1
Pillow==10.2.0
ruff==0.1.11
black==24.1.1

